requirements/development.txt
-r base.txt

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0

# Documentation
Sphinx==7.2.6